Fixed a bad bug with non-defined event handlers.
The signature check that checks if an event handler in
code matches the signature of one defined in the library data
was still being preformed even when an event signature could not be returned
from the Library Data provider object. This was causing a null object
error when trying to compile code containing an un-defined event handler.

The behavior has been fixed so that an undefined event does not get
signature checked after the undefined event error is reported, and syntax checking continues down the syntax tree into the event handler code so that other errors in the event body code can be seen.