Renamed interface ILSLMainLibraryDataProvider to ILSLLibraryDataProvider.
Added the notion of subset descriptions to ILSLLibraryDataProvider.

A subset description object should now be present for every possible
library subset that signatures in the provider may have.

If a library signature exists in the provider with a subset that lacks a
matching LSLLibrarySubsetDescription object, then it constitutes an
invalid state for the object.

Implementors of ILSLLibraryDataProvider must now insure that no library
signatures can exist in the provider that have subsets with no description
object associated to them in the provider.

Updated the editor UI to take advantage of subset descriptions in library
data providers.  The Tab Library Data menu is now generated dynamically
from
the contents of the library data provider that is used.

The lslcc_cmd project has also been updated so that it reports what
library subsets are available to use when compiling by looking at the
subset descriptions in an LSLDefaultLibraryDataProvider instance that is
created in live filtering mode.