Fixed up license distribution
LSLCCEditorInstaller now places a 'ThirdPartyLicense'
directory in the install path, as well as a license.rtf
that applies to LSLCCEditor and LibLSLCC.

The LSLCCEditor project copies the licenses it needs to distribute
for its dependencies to its build output path under the
'ThirdPartyLicenses' folder, and the WIX installer
project fetches the files from that location for its build.
The licenses LSLCCEditor requires are copied from the 'ThirdPartyLicenses'
folder in the top level directory.

The LibLSLCC project now copies the licenses it requires for
its dependencies (Antlr 4) from the 'ThirdPartyLicenses' directory
at the top level directory.  I places them into the 'ThirdPartyLicenses' directory
of the projects 'bin' folder where the binaries are built.