Fixed LibraryDataScrapingTools for now
I am probably going to make LSLLibraryDataProvider track
changes to the Subset collection's of its signature objects
and have it throw an exception if you change a signature's subset
without cloning it,  or just return a clone everytime.

Either of these changes will break the way LibraryDataScrapingTools works
and I will have to change it in a few different places.