Fixed issue with opening files through explorer
The editor was looking in its current working directory
for the library_data folder.

When you open a file thats associated to the editor through
the windows UI, the working directory is not the location
in which the application is installed.

Needed to make it look in the correct folder regardless
of the working directory.