Allow duplicate var declaration in nested scope
Previously variable names could not be re-used inside
of a nested scope when they were already defined in an outer scope.

ie:

integer x = 0;

if(1)
{
	integer x = 0;
}

Would have been considered a syntax error.

The linden compiler allows this to occur, so the syntax error
has been removed and replaced with a warning that tells you
that the previous declaration is hidden for the duration of the
nested scope.