Bugfixes to autocomplete
Autocomplete after expression-less return statements
was experiencing problems due to the auto complete parser
mistakenly thinking the cursor offset was still inside of the
previous return statements expression area.

Autocomplete for functions and other things after variable
assignments was experiencing problems inside control statements.
This was because the autocomplete parser was mistakenly deducing that the cursor
offset was between control statement keywords and braces/parenthesis in
a non codeable area of the lsl syntax.  The check to see if the cursor offset
was after an assignment expression to a variable or vector/rotation
component access was not contributing to the check for being inbetween
control statement tokens and the outside of brace/parentheses tokens as it
should have been.

Demos of the problems fixed are below:

```LSL

for /* ^(parser thinks im in an area like this) */ (i=0; i < 100; i++) /*
or here*/ {

	variable = /* but im actually typing in here */ ;

}

state_entry(){

	return;

	/* parser thinks im still writing a return expression */
}

```