Fix formatting comments before braceless statement
When the indent single statement options was turned off
in the code formatter (braceless if statements, ect...)

The formatter was incorrectly indenting bracless expressions
with comments preceding the single statement.

if(1) /*test*/ llSay(0,"test");

will now be forcefully indented as a side effect for now.

prior to this fix code such as:

if(1) //test
    llSay(0, "test");

would be reformatted as:

if(1) //test llSay(0, "test");

which results in a syntax error.