Fix improper code generation issue
Expression statements such as:

~(x=0);

-(x=0);

ect..

(unary op) (expression);

And:

(x=3);

(any expression with parenths as statement);

Were not being accounted for when compiling to CSharp.

CSharp does not allow anything but variable declarations,
variable assignments/modifications, and method calls to be used
as a standalone statement.

Therefore these valid LSL expressions must be sanitized when compiling.