Editor fixes for external file modify detection
The intial tab that gets created when the editor starts
was not being specified as selected upon loading.  leading to an
external file change check not occuring until you moved to another
tab and back.

Fixed various message prompts that should have been topmost windows but
were not.  This prevents interaction with the editor window for several
existing message prompts, such as the one that informs you a file has been
externally modified, as well as the one's that inform you that you are
overwriting editor settings, etc.

There were quite a few prompts that needed this fix.