Update version.py system
A source file that determines the embeded assembly version for each
project has been put into a folder named 'versions' in the top level
directory.

This is because the revision number for each project is based off the
number of changes git detects in the projects directory.

I do not want the revision number (change count) to increment when the
assembly version is changed in AssemblyInfo.cs, this causes a feedback loop
with the revision number because that file exists in the projects source tree.

Each project now references a Version.cs file in the 'versions' folder
outside of its source tree, so that automated changes to those files
will not cause the revision number for each assembly to increment.