Update build.py for versioning
Binary releases produced by build.py now have a version stamp
instead of a date stamp.  The library release zip file (that also includes
the command line compiler) is tagged with the library version.

The installers are tagged with LSLCCEditor's version.

When the wix installer is built through build.py,  build.py sets
the proper version definition for the installer that ends up
as part of the compiled .msi file.

If the wix installer is built without build.py, the version
on the installer defaults to 1.0.0.0.

===

Fixed CLS compliance warnings in code.

===

Changed the installer markup to allow for product downgrades.