Differentiate between ParentScopeId and ScopeId
ILSLCodeStatement now uses ParentScopeId to denote the
scope ID of the parent scope it exists in.

ILSLCodeScopeNode now uses ScopeId to denote the ID of
the code scope, which gets put into each childs ParentScopeId.

Code scopes are statements; so code scopes have both a 'ScopeId'
(the ID passed to the childrens 'ParentScopeId'), and a 'ParentScopeId'
(the ID of the parent scope, if a scope is an anonymous block)