Box key variables/casts in list literal init
Key type variables are now boxed when put into a list literal
expression, from LSL_Types.LSLString to LSL_Types.key

This also occurs for strings are cast to keys in list literal initializers.

It even works when the variable/cast is in a parenthesized expression.

The cast from string to key is usually discarded in the OpenSim compiler
because OpenSims runtime functions do not distinguish between the two.
(frustrating)

llGetListEntryType does not even test for the entry type being
LSL_Types.key in OpenSim,  it's considered a key if it can be parsed
as a UUID which is.. interesting.

This does not fix the fact that llGetListEntryType is broken, or the
fact that most OpenSim runtime functions that should return some form
of key type actually return a string.

but it fixes the type that is actually put in the list initializer
when key variables are passed in, since they have to compile to strings
to suit OpenSims runtime.

llGetListEntryType will still not report LSL_Types.key entrys as keys,
unless they are actually formatted like a key.  that is ridiculous.