Fix about window info display/loading issue
The window was attempting to load license.rtf from the current
working directory, which caused it to be missing when opening a script
using windows file association (double clicking the file)

This is because the working directory is not actually where the exe is
located in that case.

This also caused problems with displaying loaded program dependencies.