Remove bit about monodevelop from README.md
Modern monodevelop/mono cannot support building solutions targeting .NET v4.0

The LibLSLCC project solution targets v4.0 by default to stay inline
with OpenSims default target.

In order to build on mono you need to force the framework target for
LibLSLCC to v4.5

It can be done on the command line, and the python build script takes
care of it, but it is unwieldy to do from inside mono develop without
changing the solution file itself.

Removed the section about building from inside monodevelop as it is more
user friendly to just build the project from the command line.