Fix outstanding bug with tab autosave prompts on close
Fix crash by clicking 'Cancel' when a changed tab is trying to autosave
on close.

Fix behavior of clicking 'No' when a changed tab is trying to autosave
on close (IE close it and select the next changed tab)

Fix the save prompt when autosaving an in memory file, clicking cancel
in the save prompt now cancels out of the entire exit process instead of
looping and creating the window again.